Might mess everything up so checking this in
had to redo timestamps, more input validation, messages are sending now but not being receiving so need to swap out the listening server socket. Message class functionality modified to correspond to what the methods need to use it for and proper handling of it along with the timestamp from the broadcast method. Processes now include their own host in the list of hosts to build the vector clock, the hosts of the inputs are sorted so that each index actually corresponds to a process as opposed to being the order the processes connect in which previously made the current process a bit risky to enter in the list incase two processes were started simultaneously but now there is more reliance on the user