Reimplemented from the ground up
Worked to redesign the previous implementation as I felt it had become too complex and disorganized, and I was beginning to see flaws in the ways that I was connecting. I started over with a simple program that is ran on two processes, each has a server setup and sends a message which is read by the other. Then I added more processes so that it takes a list of them. Then I added the timestamp and a formal message structure. Then I added the timestamp parsing, and following that the timestamp updating. Lastly I implemented the buffering and the corresponding algorithm to ensure that the processes are truly following the Schiper-Eggli-Sandoz casually ordered broadcasting algorithm.