Total ordering implemented
Processes are restricted from sending more than one request at a time with flags to coordinate thread actions between the listening server and the broadcasting thread. The Ricart-Agrawala algorithm has been implemented with a placeholder critical section to ensure that the process will make it once for each request that it sends out. Testing and debugging are up next.