PrintWriter errors
It appears there may be some connection issues, not sure why but the server seems to love dropping them randomly. Then I found out that the print writer also loves not sending messages, so it's a bit of a war on two fronts right now to get the servers and clients to talk to each other.