Various changes for bugfixing
Current version is still a trainwreck but at least the processes are connecting, sending messages, and not reaching timestamps and critical section counts in the tens of thousand such as in various other editions of the program today. Message passing fixed, cause was a vs code suggestion for socket resource leaks causing my socket connections to immediately fall out of scope at die upon being created, awesome. Hosts and remoteHosts index discrepancy re-realized and subsequent changes applied, writers array is now parallel to the hosts with the index for the local process being null as this way is much easier for various reasons. Subsequent connection issues fixed as a result of print writer changes and host index determining fixed. Still some issues with processes seeming to send messages to themselves randomly but it's 2am and that's a problem for tomorrow (later today 😋)