Some more fixes, more on the way
More host connection issues I overlooked when I wrote my last batch of fixes at 2am. Currently, the flag for outstanding messages is turned off so that just the first batch of requests can be processed correctly. However, for some reason, the replies and requests seem to continue to crawl despite only one series of requests from each process being sent. More investigation required but at least this version has no errors. It also might be possible that the vector clocks aren't updating, but that's likely a result of the other problem, which is likely to be a result of processes retaining messages or something